reformated navigation to be one component
reformated the navigation menus to be a single component that checks the current location to choose what links to display
