+++ frontend/src/components/DashboardComp.js
@@ -2,7 +2,7 @@ import { Container, Paper } from '@mui/material';
 import { Box } from '@mui/system';
 import { useAuthorizedContext } from '../context/AuthContext';
 import { AccountOverview } from './AccountOverview';
-import { NavigationButtonsDashboard } from './NavigationButtonsDashboard';
+import { NavigationButtons } from './NavigationButtons';
 
 //the actual contents of the dashboard. The Dashboard page uses this to conditionally render the dashboard contents
 
@@ -18,7 +18,7 @@ export const DashboardComp = () => {
       <Container>
         <Paper elevation={3}>
           <Box sx={{ width: '100%', padding: '16px' }}>
-            <NavigationButtonsDashboard />
+            <NavigationButtons />
             <AccountOverview />
           </Box>
         </Paper>
