+++ frontend/src/components/SettingsComp.js
@@ -2,6 +2,7 @@ import { Container, Paper } from '@mui/material';
 import { Box } from '@mui/system';
 import { EmailChange } from './EmailChange';
 import { DeleteAccount } from './DeleteAccount';
+import { NavigationButtons } from './NavigationButtons';
 
 //the actual contents of the settings page
 
@@ -13,10 +14,11 @@ export const SettingsComp = () => {
         <Container>
           <Paper elevation={3}>
             <Box sx={{width: '100%', padding: '16px'}}>
-                <div className='d-flex flex-column align-items-center'>
-                    <EmailChange />
-                    <DeleteAccount/>
-                </div>
+              <NavigationButtons />
+              <div className='d-flex flex-column align-items-center'>
+                  <EmailChange />
+                  <DeleteAccount/>
+              </div>
             </Box>
           </Paper>
         </Container>
